refactor code and improve the test suite
