+++ lib/figo.rb
@@ -22,179 +22,16 @@
 # THE SOFTWARE.
 #
 
-require 'json'
 require 'active_support/core_ext/object/to_query'
 
-require_relative './helpers/https.rb'
-require_relative './helpers/error.rb'
-require_relative './authentification/api_call.rb'
+require_relative 'authentification/api_call'
+require_relative 'catalog/api_call'
+require_relative 'catalog/model'
+require_relative 'connection'
+require_relative 'helpers/error'
+require_relative 'helpers/https'
+require_relative 'session'
 
-# Ruby bindings for the figo Connect API: http://developer.figo.me
 module Figo
   API_ENDPOINT = 'api-preview.figo.me'
-
-  # Represents a non user-bound connection to the figo Connect API.
-  #
-  # It's main purpose is to let user login via OAuth 2.0.
-  class Connection
-    require_relative './catalog/model.rb'
-    require_relative './catalog/api_call.rb'
-
-    include Figo
-    # Create connection object with client credentials.
-    #
-    # @param client_id [String] the client ID
-    # @param client_secret [String] the client secret
-    # @param redirect_uri [String] optional redirect URI
-    def initialize(client_id, client_secret, redirect_uri = nil)
-      @client_id = client_id
-      @client_secret = client_secret
-      @redirect_uri = redirect_uri
-      @https = HTTPS.new("figo-#{client_id}", nil)
-    end
-
-    # Helper method for making a OAuth 2.0 request.
-    #
-    # @param path [String] the URL path on the server
-    # @param data [Hash] this optional object will be used as url-encoded POST content.
-    # @return [Hash] JSON response
-    def query_api(path, data = nil, method = 'POST')
-      uri = URI("https://#{API_ENDPOINT}#{path}")
-
-      # Setup HTTP request.
-      request = method == 'POST' ? Net::HTTP::Post.new(path) : Net::HTTP::Get.new(path)
-      request.basic_auth(@client_id, @client_secret)
-      request['Accept'] = 'application/json'
-      request['Content-Type'] = 'application/x-www-form-urlencoded'
-      request['User-Agent'] = 'figo-ruby/1.4.2'
-      request.body = URI.encode_www_form(data) unless data.nil?
-
-      # Send HTTP request.
-      response = @https.request(uri, request)
-
-      # Evaluate HTTP response.
-      response.body && !response.body.empty? ? JSON.parse(response.body) : nil
-    end
-
-    def query_api_object(type, path, data = nil, method = 'GET') # :nodoc:
-      response = query_api path, data, method
-      return nil if response.nil?
-
-      type.new(self, response)
-    end
-
-    def get_version
-      query_api '/version', nil, 'GET'
-    end
-  end
-
-  # Represents a user-bound connection to the figo Connect API and allows access to the user's data.
-  class Session
-    require_relative './access/model.rb'
-    require_relative './access/api_call.rb'
-
-    require_relative './account_balance/model.rb'
-    require_relative './account_balance/api_call.rb'
-
-    require_relative './account/model.rb'
-    require_relative './account/api_call.rb'
-
-    require_relative './bank/model.rb'
-    require_relative './bank/api_call.rb'
-
-    require_relative './catalog/model.rb'
-    require_relative './catalog/api_call.rb'
-
-    require_relative './notification/model.rb'
-    require_relative './notification/api_call.rb'
-
-    require_relative './payment/model.rb'
-    require_relative './payment/api_call.rb'
-
-    require_relative './strong_customer_authentication/auth_method/model.rb'
-    require_relative './strong_customer_authentication/synchronization_challenge/model.rb'
-    require_relative './strong_customer_authentication/api_call.rb'
-
-    require_relative './synchronization_status/model.rb'
-    require_relative './synchronization_status/api_call.rb'
-
-    require_relative './sync/model.rb'
-    require_relative './sync/api_call.rb'
-
-    require_relative './transaction/model.rb'
-    require_relative './transaction/api_call.rb'
-
-    require_relative './user/model.rb'
-    require_relative './user/api_call.rb'
-
-    require_relative './standing_order/model.rb'
-    require_relative './standing_order/api_call.rb'
-
-    require_relative './security/model.rb'
-    require_relative './security/api_call.rb'
-
-    require_relative './task/model.rb'
-    require_relative './task/api_call.rb'
-
-    include Figo
-
-    # Create session object with access token.
-    #
-    # @param access_token [String] the access token
-    def initialize(access_token)
-      @access_token = access_token
-      @https = HTTPS.new("figo-#{access_token}", nil)
-    end
-
-    # Helper method for making a REST request.
-    #
-    # @param path [String] the URL path on the server
-    # @param data [hash] this optional object will be used as JSON-encoded POST content.
-    # @param method [String] the HTTP method
-    # @return [Hash] JSON response
-    def query_api(path, data = nil, method = 'GET') # :nodoc:
-      uri = URI("https://#{API_ENDPOINT}#{path}")
-
-      # Setup HTTP request.
-      request = case method
-                when 'POST'
-                  Net::HTTP::Post.new(path)
-                when 'PUT'
-                  Net::HTTP::Put.new(path)
-                when 'DELETE'
-                  Net::HTTP::Delete.new(path)
-                else
-                  Net::HTTP::Get.new(path)
-                end
-
-      request['Authorization'] = "Bearer #{@access_token}"
-      request['Accept'] = 'application/json'
-      request['Content-Type'] = 'application/json'
-      request['User-Agent'] = 'figo-ruby/1.4.2'
-
-      request.body = JSON.generate(data) unless data.nil?
-
-      # Send HTTP request.
-      response = @https.request(uri, request)
-
-      # Evaluate HTTP response.
-      return nil if response.nil? || response.body.nil? || response.body.empty?
-
-      JSON.parse(response.body)
-    end
-
-    def query_api_object(type, path, data = nil, method = 'GET', array_name = nil) # :nodoc:
-      response = query_api path, data, method
-      return nil if response.nil?
-      return type.new(self, response) if array_name.nil?
-
-      response[array_name].map { |entry| type.new(self, entry) }
-    end
-
-    def encodeOptions(options)
-      options.delete_if { |k, v| v.nil? }.to_query
-      optionsEncoded = URI.encode_www_form(options)
-      return optionsEncoded
-    end
-  end
 end
